feature/frontend-error-message (#199)
* add error message in frontend

* fix codacy issue
