+++ frontend/src/api/index.js
@@ -58,41 +58,3 @@ export const getManifest = () => {
     url: '/api/manifest'
   })
 }
-
-
-/**
-    Get data list by data group name
-    {
-        "conf": {
-            "filters": [
-            {
-                "contents": [
-                "/fe"
-                ],
-                "response": "fe_c07ba164-bbd4-4f02-bd54-7d78a078a09d"
-            }
-            ],
-            "parent": null
-        },
-        "data": [
-            [
-            "fe_c07ba164-bbd4-4f02-bd54-7d78a078a09d",
-            "data/Test/fe_c07ba164-bbd4-4f02-bd54-7d78a078a09d"
-            ]
-        ]
-    }
-    */
-export const getDataList = (groupId) => {
-  return axios({
-    url: '/api/mock/' + groupId
-  })
-}
-
-
-export const getConflict = (dataId) => {
-  return axios({
-    url: '/api/conflict/id/' + dataId
-  })
-}
-
-
