+++ frontend/src/views/Main.vue
@@ -102,6 +102,11 @@ export default {
     this.$store.dispatch('loadStatus')
     this.$store.dispatch('loadManifest')
   },
+  created () {
+    this.$bus.$on('msg.success', this.successMessage)
+    this.$bus.$on('msg.info', this.infoMessage)
+    this.$bus.$on('msg.error', this.errorMessage)
+  },
   computed: {
     menuitemClasses () {
       return ["menu-item", this.isCollapsed ? "collapsed-menu" : "menu"];
@@ -147,6 +152,27 @@ export default {
       } else {
         window.open(menuItem.path, '_self');
       }
+    },
+    successMessage (msg) {
+      this.$Message.success({
+        content: msg,
+        duration: 3,
+        closable: true
+      })
+    },
+    infoMessage (msg) {
+      this.$Message.info({
+        content: msg,
+        duration: 3,
+        closable: true
+      })
+    },
+    errorMessage (msg) {
+      this.$Message.error({
+        content: msg,
+        duration: 0,
+        closable: true
+      })
     }
   }
 };
